✨ feat(pricing): update distance parameters to use Float for improved precision
- change distance parameters from Double to Float in pricing calculations
- update spawn distance handling to use Float for consistency
- introduce environment multiplier for pricing based on world type

+++ src/main/kotlin/dev/slne/surf/protect/paper/message/Messages.kt
@@ -231,7 +231,7 @@ object Messages {
             effectiveCost: Double,
             currency: Currency,
             pricePerBlock: Double,
-            distanceToSpawn: Double,
+            distanceToSpawn: Float,
             discountFactor: Double = 1.0
         ) = buildText {
             val distanceToSpawn = (distanceToSpawn * 100).roundToInt() / 100.0
